BE/requirements.txt
flask==3.0.0
pymysql==1.1.0
pydantic==2.5.2
python-dotenv==1.0.0
cryptography==42.0.5
flask-cors==5.0.0